apps/server: fix command injection on windows by resolving executables and avoiding shell: true
